# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentWarden API
  version: 0.1.0
  summary: Pre-sign EVM transaction security analysis for AI agents.
servers:
  - url: http://localhost:8787
paths:
...
  /reports/{reportHash}:
    get:
      summary: Retrieve a persisted report
      description: Returns a locally persisted transaction or signature report when REPORT_STORE_DIR is configured.
      operationId: getReport
      parameters:
        - name: reportHash
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ReportHash"
      responses:
        "200":
          description: Persisted report
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SecurityReport"
                  - $ref: "#/components/schemas/SignatureSecurityReport"
        "400":
          description: Malformed report hash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ReportHash:
      type: string
      pattern: "^0x[a-f0-9]{64}$"
    Address:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
    Hex:
      type: string
      pattern: "^0x([a-fA-F0-9]{2})*$"
      maxLength: 131074
      description: 0x-prefixed hex capped at 64 KiB of calldata.
    DecimalString:
      type: string
      pattern: "^\\d+$"
...
    SecurityReport:
      type: object
      required:
        - verdict
        - riskScore
        - summary
        - explanation
        - findings
        - recommendedAction
        - transactionEnvelope
        - actionType
        - executionGraph
        - decodedActions
        - assetDeltas
        - approvalFindings
        - decodedTransaction
        - policyViolations
        - simulationResult
        - reportHash
      properties:
        requestId:
          type: string
        verdict:
          enum:
            - ALLOW
            - WARN
            - BLOCK
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        riskVector:
          $ref: "#/components/schemas/RiskVector"
        summary:
          type: string
        explanation:
          type: string
        findings:
          type: array
          items:
            type: object
        recommendedAction:
          type: string
        transactionEnvelope:
          type: object
        actionType:
          type: string
        executionGraph:
          $ref: "#/components/schemas/ExecutionGraph"
        decodedActions:
          type: array
          items:
            type: object
        assetDeltas:
          type: array
          items:
            type: object
        approvalFindings:
          type: array
          items:
            type: object
        benchmarkProfile:
          enum:
            - agentkit
            - goat
            - eliza
            - generic
        decodedTransaction:
          type: object
        policyViolations:
          type: array
          items:
            type: object
        simulationResult:
          type: object
        stateSnapshot:
          $ref: "#/components/schemas/ChainStateSnapshot"
        stateFindings:
          type: array
          items:
            type: object
        saferAlternative:
          type: string
        reportHash:
          type: string
          pattern: "^0x[a-f0-9]{64}$"
    SignatureSecurityReport:
      type: object
      required:
        - verdict
        - riskScore
        - actionType
        - decodedSignature
        - policyViolations
        - reportHash
      properties:
        requestId:
          type: string
        verdict:
          enum:
            - ALLOW
            - WARN
            - BLOCK
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        actionType:
          type: string
        decodedSignature:
          type: object
        policyViolations:
          type: array
          items:
            type: object
        saferAlternative:
          type: string
        reportHash:
          type: string
          pattern: "^0x[a-f0-9]{64}$"
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    RiskVector:
      type: object
      required:
        - contractRisk
        - tokenRisk
        - behaviorRisk
        - intentDelta
        - sanctionsRisk
        - simulationRisk
      properties:
        contractRisk:
          type: integer
          minimum: 0
          maximum: 100
        tokenRisk:
          type: integer
          minimum: 0
          maximum: 100
        behaviorRisk:
          type: integer
          minimum: 0
          maximum: 100
        intentDelta:
          type: integer
          minimum: 0
          maximum: 100
        sanctionsRisk:
          type: integer
          minimum: 0
          maximum: 100
        simulationRisk:
          type: integer
          minimum: 0
          maximum: 100
    ExecutionGraph:
      type: object
      required:
        - rootNodeId
        - nodes
        - edges
        - maxDepth
        - hasNestedExecution
        - hasUnknownNode
      properties:
        rootNodeId:
          type: string
        nodes:
          type: array
          items:
            type: object
            required:
              - id
              - depth
              - kind
              - actionType
              - functionName
              - selector
              - warnings
              - evidence
            properties:
              id:
                type: string
              parentId:
                type: string
              depth:
                type: integer
                minimum: 0
              kind:
                enum:
                  - root
                  - nested_call
                  - multicall_child
                  - user_operation_hint
              actionType:
                type: string
              functionName:
                type: string
              selector:
                $ref: "#/components/schemas/Hex"
              evidence:
                type: array
                items:
                  type: string
              warnings:
                type: array
                items:
                  type: string
        edges:
          type: array
          items:
            type: object
            required:
              - from
              - to
              - relationship
            properties:
              from:
                type: string
              to:
                type: string
              relationship:
                enum:
                  - contains
                  - approves
                  - transfers
                  - delegates
                  - executes
                  - unknown
        maxDepth:
          type: integer
          minimum: 0
        hasNestedExecution:
          type: boolean
        hasUnknownNode:
          type: boolean
    ChainStateSnapshot:
      type: object
      required:
        - chainId
        - blockTag
        - account
        - erc20
        - erc721
        - erc1155
        - lookupErrors
      properties:
        chainId:
          type: integer
        blockTag:
          enum:
            - latest
        account:
          type: object
          properties:
            address:
              $ref: "#/components/schemas/Address"
            nativeBalance:
              $ref: "#/components/schemas/DecimalString"
            nonce:
              type: integer
        target:
          type: object
          properties:
            address:
              $ref: "#/components/schemas/Address"
            bytecode:
              $ref: "#/components/schemas/Hex"
            isContract:
              type: boolean
        erc20:
          type: array
          items:
            type: object
        erc721:
          type: array
          items:
            type: object
        erc1155:
          type: array
          items:
            type: object
        lookupErrors:
          type: array
          items:
            type: object